backend/services/brain_service/requirements.txt
ollama
websockets==12.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
fastapi
uvicorn[standard]
httpx
